# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.MFAPolicy:
    properties:
      allowed_methods:
        description: totp, sms, email
        items:
          type: string
        type: array
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      required:
        type: boolean
      updated_at:
        type: string
    type: object
...
  models.PasswordPolicy:
    properties:
      created_at:
        type: string
      history_count:
        type: integer
      id:
        type: integer
      lockout_duration:
        description: minutes
        type: integer
      lockout_threshold:
        type: integer
      max_age:
        description: days
        type: integer
      min_length:
        type: integer
      require_lowercase:
        type: boolean
      require_numbers:
        type: boolean
      require_special_chars:
        type: boolean
      require_uppercase:
        type: boolean
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@openauth.local
    name: API Support
    url: https://github.com/hanyouqing/OpenAuth/issues
  description: OpenAuth - Open-source Identity and Access Management (IAM) platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/hanyouqing/OpenAuth
  title: OpenAuth API
  version: "1.0"
paths:
  /admin/mfa-policy:
    get:
      description: Get current MFA policy configuration (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: MFA policy
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get MFA policy
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update MFA policy configuration (admin only)
      parameters:
      - description: MFA policy data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MFAPolicy'
      produces:
      - application/json
      responses:
        "200":
          description: Policy updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update MFA policy
      tags:
      - admin
  /admin/password-policy:
    get:
      description: Get current password policy configuration (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Password policy
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get password policy
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update password policy configuration (admin only)
      parameters:
      - description: Password policy data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PasswordPolicy'
      produces:
      - application/json
      responses:
        "200":
          description: Policy updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update password policy
      tags:
      - admin
...
  /api-keys:
    get:
      description: Get list of API keys, optionally filtered by user or application
        (admin only)
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by application ID
        in: query
        name: application_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: API key list
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      description: Create a new API key for user or application (admin only)
      parameters:
      - description: API key data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: API key created (key shown only once)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - api-keys
  /api-keys/{id}:
    delete:
      description: Delete API key by ID (admin only)
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: API key deleted
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - api-keys
    get:
      description: Get API key details by ID (admin only)
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: API key details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: API key not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get API key by ID
      tags:
      - api-keys
    put:
      consumes:
      - application/json
      description: Update API key information by ID (admin only)
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      - description: API key data to update
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: API key updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update API key
      tags:
      - api-keys
  /api-keys/{id}/revoke:
    post:
      description: Revoke (disable) an API key by ID (admin only)
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: API key revoked
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Revoke API key
      tags:
      - api-keys
  /applications:
    get:
      description: Get list of all applications (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Application list
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get application list
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: Create a new application (admin only)
      parameters:
      - description: Application data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Application'
      produces:
      - application/json
      responses:
        "200":
          description: Application created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create application
      tags:
      - applications
  /applications/{id}:
    delete:
      description: Delete application by ID (admin only)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Application deleted
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete application
      tags:
      - applications
    get:
      description: Get application details by ID (admin only)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Application details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Application not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get application by ID
      tags:
      - applications
    put:
      consumes:
      - application/json
      description: Update application information by ID (admin only)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: Application data to update
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Application updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update application
      tags:
      - applications
  /audit/logs:
    get:
      description: Get list of audit logs with filtering and pagination (admin only)
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by resource type
        in: query
        name: resource_type
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Start date (ISO 8601)
        in: query
        name: start_date
        type: string
      - description: End date (ISO 8601)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Audit log list
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List audit logs
      tags:
      - audit
  /audit/logs/{id}:
    get:
      description: Get audit log by ID (admin only)
      parameters:
      - description: Audit log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Audit log
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Audit log not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get audit log
      tags:
      - audit
  /audit/logs/export:
    get:
      description: Export audit logs to CSV or JSON (admin only)
      parameters:
      - default: json
        description: Export format (csv or json)
        in: query
        name: format
        type: string
      - description: Start date (ISO 8601)
        in: query
        name: start_date
        type: string
      - description: End date (ISO 8601)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Export successful
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Export audit logs
      tags:
      - audit
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email to user
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username/password and optional MFA code
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication failed
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      description: Logout current user and invalidate session
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Logout failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid refresh token
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Registration successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or user already exists
          schema:
            additionalProperties: true
            type: object
      summary: User registration
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using reset token
      parameters:
      - description: Reset token and new password
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or invalid token
          schema:
            additionalProperties: true
            type: object
      summary: Reset password
      tags:
      - auth
  /automation/executions/{id}:
    get:
      description: Get an automation execution by ID
      parameters:
      - description: Execution ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Execution
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Execution not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get execution
      tags:
      - automation
  /automation/workflows:
    get:
      description: List all automation workflows
      parameters:
      - description: Filter by enabled status
        in: query
        name: enabled
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Workflow list
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List workflows
      tags:
      - automation
    post:
      consumes:
      - application/json
      description: Create a new automation workflow
      parameters:
      - description: Workflow data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWorkflowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Workflow created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create workflow
      tags:
      - automation
...
  /cas/login:
    get:
      description: CAS (Central Authentication Service) protocol login endpoint
      parameters:
      - description: Service URL to redirect after login
        in: query
        name: service
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: Login page or redirect to service
      summary: CAS Login
      tags:
      - sso
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"